Fix group-by plan with min/max aggregates and multiple infixes (#1039) (#1039)
Summary:
Wrong query results were being returned by group by plans that contained min/max aggregates as well as multiple infixes within a group.

The problem is that the same variable `result` is reused when iterating through the infixes combinations, as well as when reporting if any keys were found in the current prefix. If the last infix happens to be not found, then the outer loop for prefixes will incorrectly assume that no keys were found under this prefix.

The fix is to record during infix iteration if at least one key was found. If so, reset result back to zero so that the current distinct prefix gets emitted.

Fixes https://github.com/facebook/mysql-5.6/issues/1036
Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1039

Differential Revision: D16567660 (https://github.com/facebook/mysql-5.6/commit/2d9301955c92844dce6cef6d4f885de24fce7f1c)

Pulled By: hermanlee

fbshipit-source-id: 9d7bf3ae4df