Better handle bad db_metadata value
Summary:
There is implicit require for db_metadata: it should be a json object , such as "{"shard": "<shard_name>", "rs": "<replicaset_id>"}"

If The db_metadata value is a json array, such as "[{"shard": "<shard_name>", "rs": "<replicaset_id>"}]", rapidjson will SIGSEGV when treating a json array as a json object and accessing members in the "object".

add checks to verify the db_metadata is a json object.

Reviewed By: abhinav04sharma, zhichengzhu

Differential Revision: D16569064

fbshipit-source-id: 61c3761713e