Fix partial trx condition when waiting for workers to finish
Summary:
After per DB commit ordering in D13086060, we buffer transactions in the SQL
thread until we find an event that contains DB information. While syncing
transactions (like DDL which need to be executed in isolation) we need to use
the `trx_queued` flag to figure out if we need to factor in a partial
transaction (i.e. a transaction  which is not fully parsed by SQL thread yet),
otherwise race conditions can occur because SQL thread will not wait for all
transactions to complete before queuing the sync transaction. In the case of DDL,
the race condition can cause the DDL to run before a conflicting trx, this will cause
a deadlock because of commit ordering or MDL locks.

Reviewed By: anirbanr-fb

Differential Revision: D16590942

fbshipit-source-id: 32ef00f1335