Fixing an illegal mem ref bug caused by large trxs
Summary:
Dependency replication uses table map events to find dependencies. The
coordinator assumes that the table map event is always available because in
steady state the worker doesn't destroy events of the transaction till it's
committed. This assumption is not true however, the worker will destroy buffered
transactions after it hits a max threshold. To fix this, we track the number of
events per trx in the coordinator and make it a sync trx when it reaches the
threshold. A sync trx is executed in isolation and hence doesn't do any
dependency calculation.

Reviewed By: anirbanr-fb

Differential Revision: D16568788

fbshipit-source-id: 2520fda08b8