slave-event converted to metadata-event
Summary:
This is a port of introducing metadata event in fb-mysql 5.6 to fb-mysql 8.0.
There are some subtle changes as compared to 5.6 implementation. Notably:
(1) 8.0 introduces the notion of a binlog event library (libbinlogevents) which
distinguishes between row events, control events etc. Metadata event is
considered as a control event. Class heirarchy is changed accordinglt in
sql/log_event.h and libbinlogevents/include/control_events.h
(2) Metadata event is considered an ignorable type as far as trx parser is
considered.

Differential Revision: D16553230 (https://github.com/facebook/mysql-5.6/commit/f06c96d8b06bb05cbdf8aa28f179ca7c89a52243)

fbshipit-source-id: 39f8bf78bbe