Port hlc implementation
Summary:
Port the entire feature of assigning HLC timestamp on master instance and
carrying it to slaves through replication.

Most of the implementation remains the same other than few nuances:
(1) 8.0 no longer reserves space in the cache for GTID. GTID is written directly
to binlog file during the flush stage of ordered commit. HLC also follows a
similar strategy. After writing GTID to binlog file, we write a Metadata event
containing HLC if it is enabled. Note that HLC is assigned when the GTID gets
generated.
(2) All tests are a direct port from 5.6 (other than minor  visual changes) and
all tests continue to work in 8.0 as-is

Reviewed By: abhinav04sharma

Differential Revision: D16553258

fbshipit-source-id: 8830a4a6900