Bug #30110851: SUBQUERY INVOLVES COUNT() AGGREGATE FUNCTION PERFORMANCE REGRESSION
Summary:
When determining whether to recalculate a materialized derived table for each
execution, the uncacheability of the _parent_ query block was used instead of
the rematerialize flag on the table (which is currently only true for
JSON_TABLE). This could cause unneeded rematerializations, especially when making
queries against non-merged views, leading to reduced performance.

Do note that there are situations where we do the opposite (materialize where a
derived table or CTE is only used once, so it should be streamed); a TODO has
been added.

(count() has nothing to do with it; the bug title is misleading.)

Backported from 8.0.19: https://github.com/mysql/mysql-server/commit/d3ce6ad

Reviewed By: lth

Differential Revision: D20454117

fbshipit-source-id: 2384bde5901