Fix keycache UBSan error
Summary:
I noticed that in Ubsan tests (during platform007 testing) that during mysqld startup UBSan is complaining about mismatching type:

```
/data/sandcastle/boxes/trunk-git-mysql/sql/keycaches.cc:48:5: runtime error: call to function free_key_cache(char const*, st_key_cache*) through pointer to incorrect function type 'void (*)(const char *, unsigned char *)'
/data/sandcastle/boxes/trunk-git-mysql/sql/keycaches.cc:114: note: free_key_cache(char const*, st_key_cache*) defined here
```

This behavior is indeed undefined as different ABI may choose to pass the structs of different size in different ways, but in most platforms we care about this is benign.

Note this happens today without platform007 - but for some reason this doesn't fail the test. Perhaps we need to enable some abort on error behavior - that's something we can investigate later.

Reviewed By: luqun

Differential Revision: D16796822

fbshipit-source-id: 02c96973bdf