support for exposing admin in user_statistics and adding user_table_statistics
Summary:
This diff has two changes
1. add 2 new columns, last_admin and last_non_admin, in information_schema.table_statistics.
They record the last time a table was used by an admin user and non admin user respectively.
The definition of admin is derived from a new global variable: admin_users_list.
This is a char variable that is empty by default and is expected to be populated by a comma
separate list of user names to treat as admin. For example, value "user1" will
lead to is_admin being populated with YES for user1 and NO for all others.
When the parameter is set empty the value of is_admin is null
2. add a new table: information_schema.user_table_statistics
The table provides a subset of columns in information_schema.table_statistics
for a tuple (user-name, table-schema, table-name, table-engine), i.e. the
activities reported in table_statistics is now broken down by user-name.
The table is populated only if user_table_stats_control (new global variable)
is set to "all".
The default value of the variable is "off". Another is "basic" which controls
populating the two new columns, last_admin and last_non_admin.

Differential Revision:
D16928167

Date:      Tue Aug 20 15:25:12 2019 -0700

On branch arcpatch-D16212461
Changes to be committed:
	modified:   mysql-test/r/information_schema-big.result
	modified:   mysql-test/r/information_schema.result
	modified:   mysql-test/r/information_schema_db.result
	modified:   mysql-test/r/mysqld--help-notwin-profiling.result
	modified:   mysql-test/r/mysqld--help-notwin.result
	modified:   mysql-test/r/mysqlshow.result
	modified:   mysql-test/r/show_table_stats_bugs.result
	new file:   mysql-test/r/user_table_stats_basic.result
	modified:   mysql-test/suite/funcs_1/r/is_columns_is.result
	modified:   mysql-test/suite/funcs_1/r/is_tables_is.result
	modified:   mysql-test/suite/perfschema/r/dml_setup_instruments.result
	modified:   mysql-test/suite/rocksdb/r/table_stats.result
	new file:   mysql-test/suite/sys_vars/r/admin_users_list_basic.result
	new file:   mysql-test/suite/sys_vars/r/user_table_stats_control_basic.result
	new file:   mysql-test/suite/sys_vars/t/admin_users_list_basic.test
	new file:   mysql-test/suite/sys_vars/t/user_table_stats_control_basic.test
	new file:   mysql-test/t/user_table_stats_basic.test
	modified:   sql/handler.cc
	modified:   sql/handler.h
	modified:   sql/mysqld.cc
	modified:   sql/mysqld.h
	modified:   sql/set_var.h
	modified:   sql/sql_base.h
	modified:   sql/sql_class.h
	modified:   sql/sql_connect.cc
	modified:   sql/sql_connect.h
	modified:   sql/sql_show.cc
	modified:   sql/structs.h
	modified:   sql/sys_vars.cc
	modified:   sql/table_stats.cc

fbshipit-source-id: 8a21b279551