Issue #790: MultiGet-based MRR, Secondary Keys
When scanning a secondary index, mrr_n_ranges is not the same
as #rowids.
Introduce mrr_n_rowids. Note that we don't have a reliable
estimate of how many rowids we will have.
(TODO: can we get rid of array-based allocations in form
   mrr_statuses = new (buf) rocksdb::Status[n_elements];
and invoke the constructors when we actually need them?)