Rotation of binlog file to go via Raft plugin
Summary:
if raft is enabled, then during binlog rotation, we mimic
what happens in ordered commit to get consensus

1. before flush
2. before commit (wait for rotate event to be committed)

this guarantees that at a rotation, all transactions are committed,
and we will never need to trim into that file.

We also added raft listener functions for binlog and relay log rotate.

Reviewed By: bhatvinay

Differential Revision: D17582137

fbshipit-source-id: 725591ecb65