Fix bypass bug in prefix decoding optimization
Summary:
This is a classic stale pointer bug discovered in data-correctness testing and we are unpacking int fields to garbage as a result.

In bypass we try to optimize prefix decoding for case like `SELECT d, c, b, a FROM t WHERE a=1 and b=2` with index being `(a, b, c, d)`. We unpack a, b and remember the location for the first time, and in future unpacking we can simply skip ahead to c, d. The problem is we remember the location of a/b in m_send_mapping with a pointer into m_row_buf (for unpacking) but that may reallocate and leading to stale pointers in m_send_mapping.

The fix is to use offset rather than pointers.

Also updated test case. But unfortunately it's hard to make sure the test actually did crash before the fix (usually it just prints out old value that are still there) - I did manually verify in debugging that it did trigger the condition.

Reviewed By: luqun

Differential Revision: D17221807

fbshipit-source-id: e24b0bf454c