handle consensus failures in ordered-commit
Summary:
First step in making step down to be online (not needing a restart).
If the leader of the group fails to determine if the trx has 'consensus'
committed (in semisync_stage of ordered_commit), then it sets
commit_consensus_error flag for all trxs in the group. The commit_stage of
ordered commit takes appropriate action based on this flag and the new sysvar
commit_consensus_error_action. The default action is to explicitly rollback the
trx in the engine and return a commit error to the client. Binlog contents of
the trxs that get rolled back are retained

Reviewed By: anirbanr-fb

Differential Revision: D17192922

fbshipit-source-id: 895b2c78c0b