fix MTS when HLC is enabled
Summary:
MTS imposes restrictions on binlog event ordering within a  trx. It works by
buffering all the events in a trx until it knows the database to which this trx
should be applied - this means it buffers gtid event, query begin event. With
HLC it should also buffer metadata event. This diff fixes the same and updates a
few debug asserts accordingly.

Reviewed By: abhinav04sharma

Differential Revision: D17333488

fbshipit-source-id: 4efec6c82b1