ability to remove gtid from logged gtids
Summary:
As part of leadership changes, we will trim binlogs/relay logs on a
running mysql instance. This needs the ability to remove gtids from executed_gtid set.
This diff adds the ability to do the same. One can enqueue an event in the raft
listener queue and the raft thread in the server will take care of removing and
updating executed_gtid on a running instance.

Reviewed By: anirbanr-fb

Differential Revision: D17291883

fbshipit-source-id: dc669e5a206