Fix float unpacking
Summary: MySQL 8.0 float/double packing algorithm (field_float::make_sort_key, field_double::make_sort_key) has changed and we need to update the unpacking algorithm to align with 8.0 version. In the future we should consider port 5.6's make_sort_key to 8.0 so that we are not dependent on future make_sort_key changes, as well as allowing cp_bin for rocksdb data.

Reviewed By: lloyd

Differential Revision: D17494855 (https://github.com/facebook/mysql-5.6/commit/c8dddcc83a90d8f6fcabb5b98eb6810a3948a567)

fbshipit-source-id: b57ca3b7f44