Replace object name (DB, table, index) and hash key in the statistics structure by ID
Summary:
The statistics structures that support table_statistics, index_statistics,
and user_table_statistics currently store names of objects such as table
DB and index, taking ~200 bytes per name.
In addition they store the hash key as a combination of db & table name.
The change in this diff is to replace
 - the object name by an ID based on a map maintained per object type
   The data returned by queries on user_table/index/table/db_statistics
   tables will remain the same (name fields)
 - the hash key by an MD5 hash computed from the previous hash key value

Reviewed By: lth

Differential Revision: D17463944

fbshipit-source-id: cafb1866cdc