Port MTS fixes when hlc is enabled
Summary:
This is a direct port of
https://github.com/facebook/mysql-5.6/commit/d6452f0288c956a841d6a5ba255fd6e8bb37f252

MTS imposes restrictions on binlog event ordering within a  trx. It works by
buffering all the events in a trx until it knows the database to which this trx
should be applied - this means it buffers gtid event, query begin event. With
HLC it should also buffer metadata event. This diff fixes the same and updates a
few debug asserts accordingly.

Differential Revision: D17415806 (https://github.com/facebook/mysql-5.6/commit/990e7b5ce8fa2bf4db4b9f9cb9d9f880a061bd42)

fbshipit-source-id: 6c6e827d9cd