Rebase SYS_VAR_INT negative values due to a MySQL 8 limitation that they were always displayed as UINT
Summary:
This is a temporary change that make the test happy. MySQL 8 has a behavioral change (bug, really) that it no longer properly displays negative values for MYSYS_VARS_INT - it is always shown as a unsigned value even though internally it properly treat it as signed in modifications so it doesn't impact any behavior other than displaying the variable.

Also the mysqld_restart.inc script now outputs #restart.

Reviewed By: zhichengzhu

Differential Revision: D17503645 (https://github.com/facebook/mysql-5.6/commit/ea45d466cc3c5528801496bb9ffd841740c1eda5)

fbshipit-source-id: 29bfacd083e