Fix corruption caused by CONCAT function
Summary: The CONCAT function has a special case for when the next item to concatenate in is a substring of the current temporary buffer.  However the check for this case is incorrect.  It checks to see if the pointer is greater or equal to the start of the temporary buffer correctly, but when checking on the end it mistakenly includes the byte after the temporary buffer.  If the next string happens to immediately follow the temporary buffer, the algorithm incorrectly enters this section of code and as a result the ouput buffer ends up with a length greater than the allocated length.  Once this happens, the next time the buffer is copied, only the allocated length is copied though the string continues to assume it had the larger length bytes.  This means that random bytes that happened to be in memory at the new location are treated as part of the string - essentially corrupting the result.

Reviewed By: yizhang82

Differential Revision: D17479735

fbshipit-source-id: 1282b1f8fb6