Fix startup mutex initialization issue
Summary:
In 8.0 mutex initialization is done through `safe_mutex_init` which asserts if you have any mutex initialized during global construction. The intention is to catch static mutex (and avoid using/init them in unspecified order), and in this case we initialize them at Rdb_thread constructor and those are global objects, so they are effectively static mutex.

The fix is to delay them until init and add proper overrides.

Reviewed By: lth

Differential Revision: D17494851

fbshipit-source-id: 5704e68f85b