Link against major.minor.patch version of libmysqlclient to avoid runtime conflict
Summary: Someone shipped a libmysqlclient.so.21 in the system but that was of older version that didn't have mysql_result_metadata (actually it did - but locally and not exported). As a result libmysql_test_api is failing to locate mysql_result_metadata at runtime. The fix is to always link against major.minor.patch version that we are building to avoid conflicts.

Reviewed By: zhichengzhu

Differential Revision: D17507965 (https://github.com/facebook/mysql-5.6/commit/069fd2f8634fbdb0edf4c16b3040889cf8566ee3)

fbshipit-source-id: 40ad18814e3