Fix bug with COM_RESET_CONNECTION not respecting CLIENT_INTERACTIVE
Summary:
The COM_RESET_CONNECTION command is not currently respecting
the CLIENT_INTERACTIVE flag. The CLIENT_INTERACTIVE flag is tied to the
connection, rather than the session. When COM_RESET_CONNECTION resets
session state, it's expected that wait_timeout will be set to interactive_timeout when CLIENT_INTERACTIVE is set.

Reviewed By: jkedgar

Differential Revision: D17579865

fbshipit-source-id: edd7ec49973