Disable test case code that use partition
Summary:
For tests that isn't about partition but uses partition, disable partition portion of the test. It might be a while until we have partition support.

Given that mysqltest now expects disabled_testcase instead of disabled_parsing, as well as a bug number, I'm making up a imaginary bug number BUG888001 for partition and use that in all disabled_testcase.

Reviewed By: lloyd

Differential Revision: D17622869 (https://github.com/facebook/mysql-5.6/commit/6d3cbb12179bbce5162085e03541b39abc0d2ebd)

fbshipit-source-id: 250ca6bb901