#1052: ha_rocksdb::records_in_range() vastly overestimates #rows
In ha_rocksdb::records_in_range, Do not adjust the key value. See
issue comments for justification.

The optimizer_loose_index_scans test was hitting this. The testcase
there has a key(b,d) and queries with WHERE b=... and d>=98 (the latter
condition not matching any rows in the table).