Issue #1044: rocksdb.ttl_primary_read_filtering fails sporadically (#1045)
Summary:
Make the test stable: after DROP TABLE, make sure the compaction is
run and finishes.

If we don't do this, the post-drop compaction may run during the next
testcase. It will cause a record from the next testcase to be compacted
away when the test logic doesn't expect it and the test will fail
Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1045

Reviewed By: yoshinorim

Differential Revision: D17255547

Pulled By: yoshinorim

fbshipit-source-id: 2c20e0fd139