Issue #1044: rocksdb.ttl_primary_read_filtering fails sporadically (#1045) (#1045)
Summary:
Make the test stable: after DROP TABLE, make sure the compaction is
run and finishes.

If we don't do this, the post-drop compaction may run during the next
testcase. It will cause a record from the next testcase to be compacted
away when the test logic doesn't expect it and the test will fail
Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1045

Differential Revision: D19786026 (https://github.com/facebook/mysql-5.6/commit/1198837f05d333ce6f8d40501a8e8401a001019d)

Pulled By: hermanlee

fbshipit-source-id: 72a403ad224