synchronous execution in raft listener thread
Summary:
Adds the ability for synchronous execution of server callbacks through raft
listener queue and raft listener thread. Also adds the ability for clients to
extract the result of the callback execution through a future/promise. Ordering
of execution is still retained through the queue interface. Ability to do
fire-and-forget execution is also retained.

The QueueElement is enhanced to have a pointer to a promise. If this pointer is
not null, then raft_listener_thread will fullfill the promise by setting the
callback result. The thread enqueing the element can wait on the corresponding
promise and will be blocked until the promise is fullfilled by the server thread
(raft listener thread). Note that it is the responsibility of the caller (the
thread enqueing a callback) to ensure that the pointer to the promise is valid
until the callback execution completes (i.e the caller should wait()/get() on
the future).

Reviewed By: anirbanr-fb

Differential Revision: D17741801

fbshipit-source-id: b75a4c39e46