Add an option to mysqlbinlog to abort if invalid events are found on in-use binlogs
Summary:
If mysqlbinlog finds a corrupted binlog event (e.g. binlog event checksum failure), it aborts if the binlog file was properly closed. However, it silently exits with exit code 0 if the corrupted event was found on the active (marked as LOG_EVENT_BINLOG_IN_USE_F) binlog files.

But in some cases, we may want to know if binlog files have corrupted events on in-use (not properly closed) binlog files as well, rather than silently ignoring errors.

Adding a new command line argument "skip-verify-if-open" to control whether to ignore event errors on in-use binlog files.

Reviewed By: yoshinorim

Differential Revision: D17861482

fbshipit-source-id: c22d046a2b3