Port PLUGIN_VAR_ALLOCATED
Summary:
This is a partial port of following patch:

https://github.com/facebook/mysql-5.6/commit/ac1a06f8df83c75c9ffd31ec4dd0e4ab5a120b2c

The original patch adds ton of historgrams for various I/O tracking which we'll port separately. This is needed to fix memory leaks detected by asan.

Also, update the value of PLUGIN_VAR_ALLOCATED as it conflicts with a new PLUGIN_VAR_PERSIST_ONLY, which would lead to failure of a few group replication tests.... (which was a bit annoying to debug). I'm making the value as high as possible (0x80000000 doesn't work as MySQL converts it to a int, which I could fix but I decided to leave it alone for now - so I settled for 0x40000000).

Reviewed By: lth

Differential Revision: D17940996

fbshipit-source-id: d7525b7e488