Remove rpl_crash_safe_gtid as it doesn't use GTID and I suppose we don't care
Summary:
The test didn't use GTID and expects crash recovery to work. The reason it worked back in 5.6 is that it set relay log repository to TABLE which means it is using MyRocks, so in the case of binlog ahead of engine log, the engine log and
master binlog position saved in relay log info table is always consistent, so replaying from that location would work correctly as long as you have idempotent recovery. While as in 8.0 the relay log info is actually in innodb and so it could be either ahead or behind rocksdb WAL and there is no guarantee it will be consistent. Given that the test is not using GTID, our GTID_EXECUTED recovery logic won't help as it doesn't reset master log position.

In theory we could've fixed this by having GTID_EXECUTED recovery reset the master binlog position as well and so it would work correctly (once we have dual engine recovery working), but there seems to be no point maintaining this scenario going forward as everything is GTID based in prod.

Reviewed By: zhichengzhu

Differential Revision: D18141522

fbshipit-source-id: 808f684ee58