handle binlog trimming across file rotation
---------- https://github.com/facebook/mysql-5.6/commit/ec039b1a46a ----------

handle binlog trimming across file rotation

Summary:
Fix 1:
If the following sequence of events happen, then binlog trimming will leave some
extraneous gtid in the binlog
1. Binlog file is rotated
2. A few transactions successfully flushes to binlog file, but not yet commited
to engine
3. mysqld crashes

On crash recovery, engine will point to the position in the old binlog file.
Which means extraneous trxs in the current binlog file will not be trimmed and
gtid_executed will contain these trxs. This is not an ideal scenario.

The diff fixes it by taking into account the file name stored in engine and the
current binlog file name being recovered.
1. If both file names are the same, then trimming is based solely on the
position.
2. If file names are different and the file stored in engine is older (as
noticed by the index of the file name), then we trim the current binlog file to
the begining position of the first gtid event i.e all transactions in the
current binlog file are trimmed.

Note that this does not handle the case of file names wrapping around and this
can be a rare scenario which can be fixed by instance replacements.

Fix 2:
Binlog rotation is considered a sync point for existing transactions to flush.
This is achieved by making binlog rotation block until all prepared
transactions are committed. However, the trxs that do not generate XID are not
considered 'prepared' and hence binlog rotation is not blocked by such
transactions.This could cause some extraneous GTID to be present in a
previously rotated binlog file which never got committed. This diff fixes the
problem by making binlog rotation to block on transactions without xid.

Originally Reviewed By: abhinav04sharma

fbshipit-source-id: c4d98f06025

---------- https://github.com/facebook/mysql-5.6/commit/9420260b936 ----------

fix failure to start

Summary:
Allow for recovery to proceed without special trimming if engine and recovering
binlog file prefix do not match

Originally Reviewed By: abhinav04sharma

fbshipit-source-id: fa9a466