fix the error code when trying to drop non-existent partitions
Summary:
Dropping a non-existent partition returns unexpected error if the number of partition arguments in the query >= number of partitions a table has. Ideally we should first check whether the partitions in the query is existing or not and then check if the number of partitions we are dropping is equal to what exists in the table.

upstream bug: https://bugs.mysql.com/bug.php?id=85238

Reviewed By: lth

Differential Revision: D18482409

fbshipit-source-id: a8bc2cd4c8e