Clearing coordinator's mem root on every end event
Summary:
Vanilla MTS clears the rli's mem root (which is just preallocated block
of memory in THD) after encountering an end event in
Log_event::get_slave_worker(). We need to do the same in dependency replication.

The coordinator's mem root is usually not used since it's used only by threads
that execute transactions. In dependency replication however, we create a
conversion table when the schema between the master and slave is different to
parse out keys correctly. This conversion table is allocated in the mem root.
Note that not clearing mem root is not really a memory leak since it will be
cleared when the coordinator is stopped but it still causes rss memory to
increase.

Reviewed By: anirbanr-fb

Differential Revision: D18479423

fbshipit-source-id: f8289664ca9