Delete all apply logs and apply index file on promotion
Summary:
When mysqld is promoted as a leader, the binlog persona is changed from
apply-log to that of binlog. At this stage we should cleanup all the apply logs
and the corresponding apply log index file. This is needed so that on a restart,
the recovery code path can make the right decision and use either the apply
binlogs or regular binlog for recovering mysqld.

Some things left as a todo:
1. Better safety around deleting the apply log files
2. Crash safety of apply binlog index file needs to be revisited
3. Better deterministic testing

Reviewed By: anirbanr-fb

Differential Revision: D18351398

fbshipit-source-id: 234e6e52919