fix check for skipping metadata log event
Summary:
rli->curr_group_seen_gtid is set only for multi threaded slaves (with dependency
repl or vanilla MTS). Hence the check for skipping metadata_log_event should
have a check for parallel or single threaded slave mode.

Currently in MTR, the slaves are always skipping metadata_log_event. This makes
the raft term and index to be uninitialized - both set to -1. Hence
waitConsensusCommit() is actually not waiting for the trx to commit and the
slave ends up commiting all trxs to engine unconditinally

Reviewed By: anirbanr-fb

Differential Revision: D18597636

fbshipit-source-id: e77f8d28514