Streaming compression
Summary:
Extend protocol to support streaming compression. This adds new possible values for the slave_compression_lib variable: zstd_stream/lz4f_stream. Also, zstd_net_compression_level and lz4f_net_compression_level can now take on negative values for even faster compression.

This reuses the existing scheme where the client requests the compression algorithm via connection attributes. If the server does not support this algorithm, the server will fall back to zlib and the client may read unexpected data (which means that all servers will have to be upgraded before it is safe for clients to request these new algorithms).

When compression fails to produce data smaller than the original data, we send data uncompressed, and we reset the context on both the compression and decompression side. In theory, we only need to do this when the compressed data is larger than 16MB because we only have 3 bytes to encode the length, but to keep things simple, we'll always reset the context when compression inflates data size.

Status counters are also added to track how frequently resets happen. Compression_context_reset counts number of resets and in combination with Compression_input_bytes, will give average amount of bytes processed before a reset happens. Compression_output_bytes is also added to track compression ratio.

This also fixes a bug in the original implementation where level was not being properly set because the #ifdef used was incorrect.

Reviewed By: abal147

Differential Revision: D18620828

fbshipit-source-id: 2cca8c95c9e