Fix async client read/write-after-free
Summary:
https://bugs.mysql.com/bug.php?id=97805

Valgrind reported read-after-free in cli_safe_read_with_ok_nonblocking.
The function was caching the NET_ASYNC pointer on the stack, but the
memory could be freed if the underlying cli_safe_read_with_ok_complete
call hit an error.

This affects the async client in 8.0.17 and should also be fixed
upstream.

Reviewed By: lth

Differential Revision: D18670770

fbshipit-source-id: 2453a57048d