Fix async client read/write-after-free
Summary:
https://bugs.mysql.com/bug.php?id=97805

Valgrind reported read-after-free in cli_safe_read_with_ok_nonblocking.
The function was caching the NET_ASYNC pointer on the stack, but the
memory could be freed if the underlying cli_safe_read_with_ok_complete
call hit an error.

This affects the async client in 8.0.17 and should also be fixed
upstream.

Reviewed By: lloyd

Differential Revision: D18670770 (https://github.com/facebook/mysql-5.6/commit/31e9b5d6403be6153d64eea4ebd8594e2d1402d5)

fbshipit-source-id: ab3b8f0ad75