revovery of mysqld in raft mode
Summary:
On startup, as part of crash recovery, mysqld does the following (among other
things):
(1) Use the binlog index file to figure out the latest binlog file which was in
use and build a list of all trxs in this binlog file
(2) Perform engine recovery and as part of it either roll forward or roll back
trxs in engine based on trxs in binlog and some startup parameters
(3) Try and trim binlogs either to get rid of half-written/corrupted txns or to
match the binlog position with the engine committed trxs. Subsequently close the
'recovered' binlog file and mark it as 'not-in-use'
(4) Calculate executed_gtid_set and lost_gtid_set by reading from the binlog
index file and the latest binlog file
(5) Open and initialize a new binlog file and update the index file using
'crash-safe-index-file'
(6) Proceed to do slave recovery (init_slave()) by using the right relay log

The diff refactors this part of the code so that it plays well with raft:
(1) If there is an 'apply' binlog index file, then use it (and consequentially
'apply' binlog files) for recovery. 'Apply' binlog files are trx logs used in a
follower (slave) instances in raft mode. Fallback to using regular binlog files
for recovery on not finding 'apply' binlog index file
(2) Based on the binlog/apply-binlog index file, get the latest transaction log
and initiate mysqld recovery
(3) Rollback all trxs during engine recovery
(4) Delegate the 'trimming' of binlog file to raft and skip mutating the
latest binlog file (including not marking it closed)
(5) Calculate executed_gtid_set and lost_gtid_set by reading the
binlog/apply-binlog index file and the latest transaction log. Read the latest
transaction log only till the file position obtained during engine recovery -
this ensures that gtid_executed matches with what is committed in the engine
(6) Open the latest binlog/apply-binlog file (instead of a new file) and use it
to initialize mysql_bin_log
(7) In slave recovery, always make the relay_log point to the binlog (i.e relay
log always points to the raft log) and skip any additional rotations of the binlog

Few things to be taken care of in subsequent diff:
(1) On promotions, get rid of all 'apply' binlogs and the 'apply binlog' index
files
(2) Crash safety of apply-binlog index file needs to be revisited
(3) Deterministic testing

Reviewed By: anirbanr-fb

Differential Revision: D18306445

fbshipit-source-id: a78b58f8866