Update lock.cc's LONG_TIMEOUT to LONG_TIMEOUT_NSEC
Summary:
main.mdl_sync was failing on valgrind because of a short timeout when
locking the table. One of the calls in lock.cc was using LONG_TIMEOUT
instead of LONG_TIMEOUT_NSEC for the nsec lock interface and timing out
before the wait condition could be checked.

Squash with D14349342

Reviewed By: luqun

Differential Revision: D18741122

fbshipit-source-id: cfc77bffb01