Fix main.ssl-big-packet
Summary:
https://bugs.mysql.com/bug.php?id=97853

The main.ssl-big-packet test was failing because during SSL, it's possible for the network socket to do writes during SSL_read (and vice versa). Hence, when we call poll, we should be listening for both in and out when doing reads and writes.

In theory, this is only true for SSL, but there's no easy way to figure out from the vio if we're doing ssl or not.

Differential Revision: D18780934 (https://github.com/facebook/mysql-5.6/commit/c532f04f4ea6eabeca4d7717a0527bd3281c1fdb)

fbshipit-source-id: 44582a122f5