Return session track information in same result
Summary:
Previously we only returned session state changes
on the final statement of a query. In the multiquery case,
if an intermediate query errors out, we don't get to see
the session change flag. Send the flag with the same
statement that first triggered the change to avoid this case

verified that the other uses of update server status are:
replication
stored procedures

It should be fine that these people are also updating the session
info

Upstream is tracking a request to add this to the upstream protocol at https://bugs.mysql.com/bug.php?id=97885

squash with D18851147

Reviewed By: lth

Differential Revision: D18866865

fbshipit-source-id: 882426d29e4